refactor(design): add public_api file for design/core
