+++ libs/design/src/core/prefix-suffix/prefixable/prefixable.ts
@@ -1,7 +1,6 @@
 import { ContentChild } from '@angular/core';
 import { DaffPrefixDirective } from '../prefix.directive';
-import { Constructor } from '../../constructor';
-
+import { Constructor } from '../../../core/public_api';
 
 /**
  * A mixin for giving a component the ability to place content before another piece of content.
