+++ libs/design/src/core/public_api.ts
@@ -1,0 +1,5 @@
+// Core
+export * from './prefix-suffix/public_api';
+export * from './colorable/colorable';
+export * from './breakpoints/breakpoints';
+export * from './constructor/constructor';
